Fix for issue #1 - cache misses when requiring
The issue is completely described in github. I added a small check to make sure that, when the platform is Win32 and both the enroute script and the handler script lies on the same drive letter, then the letter casing is consistent (that is, enroute requires the handler script with the same casing as it was required).