fix: don't obscure errors when multiple routes fail to import
The newer version of `vasync` (upgraded in this lib in 6.1.1) has a worse
error message when a pipeline callback is invoked multiple times. Now it
gives the obscure `AssertionError: idx should be equal to ndone`.

The reason the pipeline callback was being called multiple times was due
to `install` logic doing route handler imports in a `lodash.forEach`
loop. The logic would catch the exception, then `return cb1(err)`, but
the return was only for the current `forEach` iteration, so was
equivalent to a `continue` instead of terminating with an error.

So if multiple route source files failed to import, we would end up
calling the pipeline callback with an error multiple times. The `vasync`
assertion error would then obscure the original error, making these much
harder to debug. We'd commonly see this if multiple route handlers both
depended on some common helper file which was missing or had an error in
it.

This replaces the `forEach` loops with `for ... of` loops, so that when
it returns and calls the pipeline callback, it now will terminate the
loop.

This also fixes some tests that were not testing what they claimed to.
The tests were failing because the config was malformed, not because of
the intended error states. I deleted one test because this lib actually
does not even throw an error in that case.